requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
transformers>=4.35.0
torch>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
scikit-learn>=1.3.0
openai-whisper>=20231117
audio-recorder-streamlit>=0.0.8
pydub>=0.25.1
